docker/Scraper/requirements.txt
dataclasses==0.6
bs4==0.0.1
requests==2.22.0
configparser==5.0.2
psycopg2-binary>=2.8
lxml


